Refactor decimal price properties to float in models and DTOs
The code changes modify the `Product`, `ProductDto`, `CartItemDto`, and `AddProductModel` classes to change the `Price` property type from `decimal` to `float`. This change is made to align the data type of the `Price` property across these classes.

This commit message follows the established convention of providing a concise summary of the changes made, starting with an action verb (`Refactor`) to indicate the type of change.

+++ Backend/Dto/CartItemDto.cs
@@ -5,7 +5,7 @@ public class CartItemDto
     public required bool IsEvent { get; set; }
     public required long ItemId { get; set; }
     public required string Name { get; set; }
-    public required decimal Price { get; set; }
+    public required float Price { get; set; }
     public required int Quantity { get; set; }
     public required string Image { get; set; }
     public bool IsOnSale { get; set; } = false;
